requirements.txt
# Generated via uv pip compile
flask==3.0.3
annotated-types==0.6.0
anyio==4.3.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.14.0
pydantic==2.6.4
pydantic-core==2.16.3
python-dotenv==1.0.1
sniffio==1.3.1
typing-extensions==4.10.0
